% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{nla_compile}
\alias{nla_compile}
\title{Compile NLA data for on-disk storage}
\usage{
nla_compile(year, use_rappdirs, local_path)
}
\arguments{
\item{year}{numeric choice of 2007 or 2012.}

\item{use_rappdirs}{logical write files to operating system data directories at the location returned by \code{\link[rappdirs]{user_data_dir}}.}

\item{local_path}{folder path to raw downloads from `nla_get`}
}
\value{
rds
}
\description{
Compile NLA data for on-disk storage
}
