% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nla_get}
\alias{nla_get}
\title{Download NLA data from USEPA}
\usage{
nla_get(year, use_rappdirs = FALSE)
}
\arguments{
\item{year}{numeric choice of 2007 or 2012.}

\item{use_rappdirs}{logical write files to operating system data directories at the location returned by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieves NLA flat files.
}
\examples{
\donttest{
nla_get(2012)
}
}
