% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlaR-package.R
\docType{package}
\name{nlaR-package}
\alias{nla}
\alias{nlaR-package}
\title{R interface to the National Lakes Assessment}
\description{
R interface to the National Lakes Assessment
}
\author{
\email{stachel2@msu.edu}
}

