% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{nla_compile}
\alias{nla_compile}
\title{nla_compile}
\usage{
nla_compile(version = "2007")
}
\arguments{
\item{version}{optional version id}
}
\value{
rds
}
\description{
nla_compile
}
\examples{
\dontrun{
nla_compile()
}
}

