% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{nla_compile}
\alias{nla_compile}
\title{nla_compile}
\usage{
nla_compile(year)
}
\arguments{
\item{year}{numeric choice of 2007 or 2012.}
}
\value{
rds
}
\description{
nla_compile
}
\examples{
\dontrun{
nla_compile(2012)
}
}
