% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nla_get}
\alias{nla_get}
\title{nla_get}
\usage{
nla_get(dest_folder = NA)
}
\arguments{
\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieves NLA flat files.
}
\examples{
\dontrun{
nla_get()
}
}

