% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nla_load}
\alias{nla_load}
\title{nla_load}
\usage{
nla_load(year)
}
\arguments{
\item{year}{numeric choice of 2007 or 2012.}
}
\value{
list
}
\description{
nla_load
}
\examples{
\dontrun{
dt <- nla_load(2012)
}
}

